% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_large_trees.R
\name{get_tree_bee_n100}
\alias{get_tree_bee_n100}
\title{Get 100 randomly selected mega-trees of Bees from Bee Tree of Life}
\source{
\url{http://beetreeoflife.org}
}
\usage{
get_tree_bee_n100(force = FALSE)
}
\arguments{
\item{force}{Logical. Re-download even if a local cache exists. Default FALSE.}
}
\value{
A \code{multiPhylo} object. Each phylogeny also has a data frame
  \code{genus_family_root} with root node information for every unique genus
  and family, usable for inserting new tips onto the phylogeny.
}
\description{
Because of the large file size (~5 MB), this dataset is not bundled with
the package and is downloaded and cached locally on first use.
}
\references{
Henríquez-Piskulich, P., Hugall, A. F., & Stuart-Fox, D. (2023).
  A supermatrix phylogeny of the world's bees (Hymenoptera: Anthophila).
  bioRxiv 2023.06.16.545281. \doi{10.1101/2023.06.16.545281}
}
\seealso{
\code{\link{tree_bee}}
}
