% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_large_trees.R
\name{get_tree_bird_n100}
\alias{get_tree_bird_n100}
\title{Get 100 randomly selected mega-trees of Birds from Bird Tree}
\source{
\url{https://data.vertlife.org/}
}
\usage{
get_tree_bird_n100(force = FALSE)
}
\arguments{
\item{force}{Logical. Re-download even if a local cache exists. Default FALSE.}
}
\value{
A \code{multiPhylo} object. Each phylogeny also has a data frame
  \code{genus_family_root} with root node information for every unique genus
  and family, usable for inserting new tips onto the phylogeny.
}
\description{
The first 50 phylogenies use the Ericson backbone and the second 50 use the
Hackett backbone (Jetz et al. 2012). Because of the large file size (~18 MB),
this dataset is not bundled with the package and is downloaded and cached
locally on first use.
}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers,
  A. O. (2012). The global diversity of birds in space and time. Nature,
  491(7424), 444.
}
\seealso{
\code{\link{tree_bird_McTavish}}
}
