% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_large_trees.R
\name{get_tree_mammal_n100_phylacine}
\alias{get_tree_mammal_n100_phylacine}
\title{Get 100 randomly selected mega-trees of Mammals from PHYLACINE V1.2}
\source{
\url{https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/Phylogenies/Complete_phylogeny.nex}
}
\usage{
get_tree_mammal_n100_phylacine(force = FALSE)
}
\arguments{
\item{force}{Logical. Re-download even if a local cache exists. Default FALSE.}
}
\value{
A \code{multiPhylo} object. Each phylogeny also has a data frame
  \code{genus_family_root} with root node information for every unique genus
  and family, usable for inserting new tips onto the phylogeny.
}
\description{
These 100 phylogenies are randomly selected from the 1000 trees generated
by Faurby et al. (2018). Because of the large file size (~6 MB), this
dataset is not bundled with the package and is downloaded and cached locally
on first use.
}
\references{
Faurby, S., Davis, M., Pedersen, R. Ø., Schowanek, S. D.,
  Antonelli, A., & Svenning, J. C. (2018). PHYLACINE 1.2: The phylogenetic
  atlas of mammal macroecology. Ecology, 99(11), 2626-2626.
}
\seealso{
\code{\link[megatrees]{get_tree_mammal_n100_vertlife}}
}
