% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_large_trees.R
\name{get_tree_plant_n100_Carruthers}
\alias{get_tree_plant_n100_Carruthers}
\title{Get 100 randomly selected plant mega-trees (Carruthers et al.)}
\source{
\url{https://www.biorxiv.org/content/10.64898/2026.01.06.695000v1}
}
\usage{
get_tree_plant_n100_Carruthers(force = FALSE)
}
\arguments{
\item{force}{Logical. Re-download even if a local cache exists. Default FALSE.}
}
\value{
A \code{multiPhylo} object. Each phylogeny also has a data frame
  \code{genus_family_root} with root node information for every unique genus
  and family, which can be used to insert new tips onto the phylogeny later.
}
\description{
This tree was based on Carruthers et al. (2026), which in turn was an update
of the Smith and Brown (2018). We randomly selected 100 plant mega-trees from this dataset.
Because of the large file size (~135 MB), this dataset is not bundled with
the package and is downloaded and cached locally on first use.
}
\references{
Carruthers et al., (2026). A large phylogenetic tree for
  euphyllophytes. bioRxiv.
}
\seealso{
\code{\link{tree_plant_Carruthers}}
}
