% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_large_trees.R
\name{get_tree_reptile_n100}
\alias{get_tree_reptile_n100}
\title{Get 100 randomly selected mega-trees of Reptiles (Squamates) from VertLife}
\source{
\url{https://vertlife.org/data/squamates/}
}
\usage{
get_tree_reptile_n100(force = FALSE)
}
\arguments{
\item{force}{Logical. Re-download even if a local cache exists. Default FALSE.}
}
\value{
A \code{multiPhylo} object. Each phylogeny also has a data frame
  \code{genus_family_root} with root node information for every unique genus
  and family, usable for inserting new tips onto the phylogeny.
}
\description{
Because of the large file size (~15 MB), this dataset is not bundled with
the package and is downloaded and cached locally on first use.
}
\references{
Tonini, J. F. R., Beard, K. H., Ferreira, R. B., Jetz, W., &
  Pyron, R. A. (2016). Fully-sampled phylogenies of squamates reveal
  evolutionary patterns in threat status. Biological Conservation, 204, 23-31.
}
