% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_amphibian_n100}
\alias{tree_amphibian_n100}
\title{100 randomly selected Mega-tree of Amphibians from VertLife.}
\format{
A phylogeny with class "multiPhylo". It is also a list, with the first to be the "consensus" tree (the other 99 are posterior phylogenies). Compare with a normal phylo object, each phylogeny has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
<http://vertlife.org/data/amphibians/>
}
\usage{
tree_amphibian_n100
}
\description{
100 randomly selected Mega-tree of Amphibians from VertLife.
}
\references{
Jetz, W., & Pyron, R. A. (2018). The interplay of past diversification and evolutionary isolation with present imperilment across the amphibian tree of life. Nature ecology & evolution, 2(5), 850-858.
}
\keyword{datasets}
