% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_bee}
\alias{tree_bee}
\title{The Maximum Likely Mega-tree of Bees from Bee Tree of Life}
\format{
A maximum likely phylogeny with class "phylo". Compare with a normal phylo object, this phylogeny has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
http://beetreeoflife.org
}
\usage{
tree_bee
}
\description{
The Maximum Likely Mega-tree of Bees from Bee Tree of Life
}
\references{
Henríquez-Piskulich, P.; Hugall, A.F.; Stuart-Fox; D. 2023. A supermatrix phylogeny of the world’s bees (Hymenoptera: Anthophila). bioRxiv 2023.06.16.545281. doi.org/10.1101/2023.06.16.545281.
}
\keyword{datasets}
