% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_bird_n100}
\alias{tree_bird_n100}
\title{100 randomly selected Mega-tree of Birds from Bird Tree.}
\format{
A phylogeny with class "multiPhylo". It is also a list. Compare with a normal phylo object, each phylogeny has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
<https://data.vertlife.org/>
}
\usage{
tree_bird_n100
}
\description{
The first 50 phylogenies are randomly selected from those 10000 trees with Ericson backbond generated by Jetz et al. (2012); and the second 50 phylogenies are randomly selected from those 10000 trees with Hackett backbone. Node labels were added for each phylogeny.
}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O. (2012). The global diversity of birds in space and time. Nature, 491(7424), 444.
}
\keyword{datasets}
