% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_butterfly}
\alias{tree_butterfly}
\title{Tree of 2244 butterfly species}
\format{
A phylogeny with class "phylo". It is also a list. Compare with a normal phylo object, it has another data frame `tree_butterfly$genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
https://springernature.figshare.com/articles/dataset/A_global_phylogeny_of_butterflies_reveals_their_evolutionary_history_ancestral_host_plants_and_biogeographic_origins/21774899?file=39124943
}
\usage{
tree_butterfly
}
\description{
This tree was generated by Kawahara et al. (2023).
}
\references{
Kawahara, Akito Y., et al. "A global phylogeny of butterflies reveals their evolutionary history, ancestral hosts and biogeographic origins." Nature ecology & evolution 7.6 (2023): 903-913.
}
\keyword{datasets}
