% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_fish_32k_n100}
\alias{tree_fish_32k_n100}
\title{100 Mega-tree of 31516 Fish from The Fish Tree of Life}
\format{
A phylogeny with class "multiPhylo". It is also a list. Each phylogeny also has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
https://fishtreeoflife.org/downloads/actinopt_full.trees.xz
}
\usage{
tree_fish_32k_n100
}
\description{
This tree was generated by Rabosky et al. (2018). It has 31516 tips, with species missing sequence data added using birth-death models. Therefore, there are 100 such posterior distribution phylogenies to account for uncertainty. The authors suggested not to use them to conduct trait evolution analyses.
}
\references{
Rabosky, D. L., Chang, J., Title, P. O., Cowman, P. F., Sallan, L., Friedman, M., ... & Alfaro, M. E. (2018). An inverse latitudinal gradient in speciation rate for marine fishes. Nature, 559(7714), 392.
}
\seealso{
tree_fish_12k
}
\keyword{datasets}
