% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_mammal_n100_phylacine}
\alias{tree_mammal_n100_phylacine}
\title{100 randomly selected Mega-tree of Mammals from PHYLACINE V1.2.}
\format{
A phylogeny with class "multiPhylo". It is also a list. Compare with a normal phylo object, each phylogeny has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
https://github.com/MegaPast2Future/PHYLACINE_1.2/blob/master/Data/Phylogenies/Complete_phylogeny.nex
}
\usage{
tree_mammal_n100_phylacine
}
\description{
These 100 phylogenies are randomly selected from the 1000 trees generated by Faurby et al. (2018). Node labels were added here.
}
\references{
Faurby, S., Davis, M., Pedersen, R. Ø., Schowanek, S. D., Antonelli, A., & Svenning, J. C. (2018). PHYLACINE 1.2: The phylogenetic atlas of mammal macroecology. Ecology, 99(11), 2626-2626.
}
\seealso{
tree_mammal_n100_vertlife
}
\keyword{datasets}
