% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_mammal_n100_vertlife}
\alias{tree_mammal_n100_vertlife}
\title{100 randomly selected Mega-tree of Mammals from VertLife.}
\format{
A phylogeny with class "multiPhylo". It is also a list. Compare with a normal phylo object, each phylogeny has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
https://vertlife.org/data/mammals/
}
\usage{
tree_mammal_n100_vertlife
}
\description{
The first 50 phylogenies are randomly selected from those 10000 trees with Node Dating Exponential backbone provided by Upham et al. (2019); and the second 50 phylogenies are randomly selected from those 10000 trees with Fossil Birth Death Zhou et al.'s backbone. Node labels were added for each phylogeny.
}
\references{
Upham, N. S., Esselstyn, J. A., & Jetz, W. (2019). Inferring the mammal tree: species-level sets of phylogenies for questions in ecology, evolution, and conservation. PLoS biology, 17(12), e3000494.
}
\keyword{datasets}
