% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_plant_Carruthers}
\alias{tree_plant_Carruthers}
\title{Mega-tree of Plants based on Carruthers et al., 2026}
\format{
A phylogeny with class "phylo". It is also a list. Compare with a normal phylo object, it has another data frame `tree_plant_Carruthers$genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
https://www.biorxiv.org/content/10.64898/2026.01.06.695000v1
}
\usage{
tree_plant_Carruthers
}
\description{
This tree was based on Carruthers et al. (2026), which in turn was an update of the Smith and Brown (2018).
}
\references{
Carruthers et al., (2026). A large phylogenetic tree for euphyllophytes. bioRxiv.
}
\keyword{datasets}
