% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_reptile_n100}
\alias{tree_reptile_n100}
\title{100 randomly selected Mega-tree of Reptiles (Squamates) from VertLife.}
\format{
A phylogeny with class "multiPhylo". It is also a list. Compare with a normal phylo object, each phylogeny has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
https://vertlife.org/data/squamates/
}
\usage{
tree_reptile_n100
}
\description{
100 randomly selected Mega-tree of Reptiles (Squamates) from VertLife.
}
\references{
Tonini, J. F. R., Beard, K. H., Ferreira, R. B., Jetz, W., & Pyron, R. A. (2016). Fully-sampled phylogenies of squamates reveal evolutionary patterns in threat status. Biological Conservation, 204, 23-31.
}
\keyword{datasets}
