% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_shark_ray_n100}
\alias{tree_shark_ray_n100}
\title{100 randomly selected Mega-tree of Sharks, Rays, and Chimaeras from VertLife.}
\format{
A phylogeny with class "multiPhylo". It is also a list. Compare with a normal phylo object, each phylogeny has another data frame `genus_family_root`, which provides the root nodes information for every unique genus and family in the phylogeny. Such information can be used to insert new tips onto the phylogeny later.
}
\source{
https://vertlife.org/data/sharks//
}
\usage{
tree_shark_ray_n100
}
\description{
100 randomly selected Mega-tree of Sharks, Rays, and Chimaeras from VertLife.
}
\references{
Stein, R. W., Mull, C. G., Kuhn, T. S., Aschliman, N. C., Davidson, L. N., Joy, J. B., ... & Mooers, A. O. (2018). Global priorities for conserving the evolutionary history of sharks, rays and chimaeras. Nature ecology & evolution, 2(2), 288-298.
}
\keyword{datasets}
